admin_panel/requirements.txt
# Flask and extensions
Flask==2.2.3
Flask-WTF==1.1.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Bcrypt==1.0.1
Flask-Paginate==2022.1.8

# Database
SQLAlchemy==1.4.46
psycopg2-binary==2.9.5

# Data visualization
plotly==5.14.1
pandas==2.0.1

# Utilities
python-dotenv==1.0.0
Werkzeug==2.2.3
email-validator==2.0.0
Pillow==9.5.0

# Testing
pytest==7.3.1
